1. Adapt the simulator to a new adaptation layer; 2. Fix the compilation warning of ios, and the implicit conversion failed; 3. Fix the mac compilation error, due to the problem of adding the link identification -all_load;

+++ cocos/platform/ios/IOSPlatform.mm
@@ -72,7 +72,7 @@ extern int cocos_main(int argc, const char** argv);
 }
 
 - (int)getFPS {
-    return _displayLink.preferredFramesPerSecond;
+    return (int)_displayLink.preferredFramesPerSecond;
 }
 
 - (void)renderScene:(id)sender {
